LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flex-aspect-ratio-img-column-011-expected.txt


PASS .flexbox 1
PASS .flexbox 2
PASS .flexbox 3
PASS .flexbox 4
PASS .flexbox 5
PASS .flexbox 6
FAIL .flexbox 7 assert_equals:
<div class="flexbox column" style="height: 10px; width: 50px;" data-expected-height="10">
    <!-- should use min(specified=100, content height=50) = 50px as minimum height,
         which the image will shrink to due to default flex-shrink. -->
    <img src="support/10x10-green.png" style="height: 100px;" data-expected-height="50">
</div>
height expected 50 but got 10
PASS .flexbox 8
PASS .flexbox 9
PASS .flexbox 10
